requirements.txt
fastapi
uvicorn
streamlit
openai
pydantic
scikit-learn
numpy
python-dotenv